feat : ProjectPost 컴포넌트 생성 후 페이지 컴포넌트와 분리
